requirements.txt
aws-cdk-lib>=2.180.0
constructs>=10.0.0
aws-cdk.lambda-layer-kubectl-v32>=2.0.0
